pyproject.toml
[project]
name = "fastapi-my-template-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "black>=25.12.0",
    "fastapi>=0.125.0",
    "fastapi-cache2>=0.2.2",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
